feat: add support for installing to directories (#21)
* Add support for installing to directories

If DESTINATION is set as a directory, we should install the asset to the
directory with the asset name. This change adds support for such operations. We
also add a test case for installing to directories.

Tested:
  make test
  make e2e

Fixes jooola/gh-release-install#20

* Convert destination attribute to the asset file path

If DESTINATION is a directory, then `{destination}` used in the version file
should include the asset name. This change updates the logic for
`self.destination` so it always points to the asset file path.

Tested:
  make test
  make e2e

jooola/gh-release-install#20